LayoutTests/http/tests/security/sandboxed-iframe-javascript-self-navigation-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Tests that an iframe without "allow-scripts" can not navigate itself to a javascript URL.
